refactor: extract reusable base level scene
